Add files via upload
Some bugs fixed when opening the plugin. Still have to fix the sync between sound and interface (there is a bit of delay).

+++ EuclideanRhythm.cpp
@@ -27,7 +27,7 @@ EuclideanRhythm::EuclideanRhythm()
 }
 
 EuclideanRhythm::EuclideanRhythm(float rate, int bpmParam, int steps, int events,
-    int rotationParam, int velocityParam, int gateParam, int noteNumberParam, int figureStepParam, 
+    int rotationParam, int velocityParam, int gateParam, int noteNumberParam, float figureStepParam, 
     bool directionParam, bool reverseParam, bool pingPongParam, bool dottedNotesParam, bool tripletsParam,
     int channel) {
